pyproject.toml
[tool.poetry]
name = "target-bot"
version = "0.1.0"
description = ""
authors = ["vezzalinistefano <ste.vezzalini@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.5"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
